refactor: separate the GetSupportedTokens method and related tests to separate files
