+++ paymaster/paymaster.go
@@ -64,23 +64,6 @@ func NewPaymasterClient(url string, options ...client.ClientOption) (*Paymaster,
 	return paymaster, nil
 }
 
-// Get a list of the tokens that the paymaster supports, together with their prices in STRK
-//
-// Parameters:
-//   - ctx: The context.Context object for controlling the function call
-//
-// Returns:
-//   - []TokenData: An array of token data
-//   - error: An error if any
-func (p *Paymaster) GetSupportedTokens(ctx context.Context) ([]TokenData, error) {
-	var response []TokenData
-	if err := p.c.CallContextWithSliceArgs(ctx, &response, "paymaster_getSupportedTokens"); err != nil {
-		return nil, err
-	}
-
-	return response, nil
-}
-
 // TrackingIdToLatestHash gets the latest transaction hash and status for a given tracking ID.
 // Returns a TrackingIdResponse.
 //
